+++ tests/test_sql.py
@@ -3,6 +3,7 @@ import pyexcel as pe
 import datetime
 from textwrap import dedent
 from db import Session, Base, Pyexcel, engine
+from nose.tools import eq_, raises
 
 
 class TestSQL(TestCase):
@@ -59,3 +60,33 @@ class TestEmptyTable:
     def test_empty_table(self):
         sheet = pe.load_from_sql(Session(), Pyexcel)
         assert sheet is not None
+
+    @raises(Exception)
+    def test_save_as(self):
+        data = [
+            ["birth", 'name', 'weight'],
+            [datetime.date(2017, 1, 11), 'Adam', 3.0]
+        ]
+        session = Session()
+        pe.save_as(array=data, dest_session=session, dest_table=Pyexcel)
+        sheet = pe.get_sheet(session=session, table=Pyexcel)
+        eq_(str(sheet), '')
+
+    def test_save_as_1(self):
+        data = [
+            ["birth", 'name', 'weight'],
+            [datetime.date(2017, 1, 11), 'Adam', 3.0]
+        ]
+        session = Session()
+        pe.save_as(array=data, name_columns_by_row=0,
+                   dest_session=session, dest_table=Pyexcel)
+        sheet = pe.get_sheet(session=session, table=Pyexcel)
+        print(sheet)
+        content = dedent("""
+        pyexcel:
+        +------------+----+------+--------+
+        | birth      | id | name | weight |
+        +------------+----+------+--------+
+        | 2017-01-11 | 1  | Adam | 3.0    |
+        +------------+----+------+--------+""").strip('\n')
+        eq_(str(sheet), content)